Refactor capital summary section code
