+++ app/presenters/summary/sections/investments.rb
@@ -1,12 +1,8 @@
 module Summary
   module Sections
-    class Investments < Sections::BaseSection
-      def show?
-        capital && requires_full_capital
-      end
-
+    class Investments < Sections::CapitalLoopBase
       def answers # rubocop:disable Metrics/MethodLength
-        if investments.empty?
+        if records.empty?
           [
             Components::ValueAnswer.new(
               :has_investments, 'no',
@@ -15,7 +11,7 @@ module Summary
           ]
         else
           Summary::Components::GroupedList.new(
-            items: investments,
+            items: records,
             group_by: :investment_type,
             item_component: Summary::Components::Investment,
             show_actions: editable?,
@@ -24,28 +20,10 @@ module Summary
         end
       end
 
-      def list?
-        return false if investments.empty?
-
-        true
-      end
-
       private
 
-      def investments
-        @investments ||= crime_application.investments
-      end
-
-      def capital
-        @capital ||= crime_application.capital
-      end
-
-      def requires_full_capital
-        [
-          CaseType::EITHER_WAY.to_s,
-          CaseType::INDICTABLE.to_s,
-          CaseType::ALREADY_IN_CROWN_COURT.to_s
-        ].include?(crime_application.case.case_type)
+      def records
+        @records ||= crime_application.investments
       end
     end
   end
