+++ app/presenters/summary/sections/properties.rb
@@ -1,12 +1,8 @@
 module Summary
   module Sections
-    class Properties < Sections::BaseSection
-      def show?
-        capital && requires_full_capital
-      end
-
+    class Properties < Sections::CapitalLoopBase
       def answers # rubocop:disable Metrics/MethodLength
-        if properties.empty?
+        if records.empty?
           [
             Components::ValueAnswer.new(
               :has_assets, 'no',
@@ -15,7 +11,7 @@ module Summary
           ]
         else
           Summary::Components::GroupedList.new(
-            items: properties,
+            items: records,
             group_by: :property_type,
             item_component: Summary::Components::Property,
             show_actions: editable?,
@@ -24,28 +20,10 @@ module Summary
         end
       end
 
-      def list?
-        return false if properties.empty?
-
-        true
-      end
-
       private
 
-      def properties
-        @properties ||= crime_application.properties
-      end
-
-      def capital
-        @capital ||= crime_application.capital
-      end
-
-      def requires_full_capital
-        [
-          CaseType::EITHER_WAY.to_s,
-          CaseType::INDICTABLE.to_s,
-          CaseType::ALREADY_IN_CROWN_COURT.to_s
-        ].include?(crime_application.case.case_type)
+      def records
+        @records ||= crime_application.properties
       end
     end
   end
