+++ app/presenters/summary/sections/savings.rb
@@ -1,12 +1,8 @@
 module Summary
   module Sections
-    class Savings < Sections::BaseSection
-      def show?
-        capital && requires_full_capital
-      end
-
+    class Savings < Sections::CapitalLoopBase
       def answers # rubocop:disable Metrics/MethodLength
-        if savings.empty?
+        if records.empty?
           [
             Components::ValueAnswer.new(
               :has_capital_savings, 'no',
@@ -15,7 +11,7 @@ module Summary
           ]
         else
           Summary::Components::GroupedList.new(
-            items: savings,
+            items: records,
             group_by: :saving_type,
             item_component: Summary::Components::Saving,
             show_actions: editable?,
@@ -24,28 +20,10 @@ module Summary
         end
       end
 
-      def list?
-        return false if savings.empty?
-
-        true
-      end
-
       private
 
-      def savings
-        @savings ||= crime_application.savings
-      end
-
-      def capital
-        @capital ||= crime_application.capital
-      end
-
-      def requires_full_capital
-        [
-          CaseType::EITHER_WAY.to_s,
-          CaseType::INDICTABLE.to_s,
-          CaseType::ALREADY_IN_CROWN_COURT.to_s
-        ].include?(crime_application.case.case_type)
+      def records
+        @records ||= crime_application.savings
       end
     end
   end
